requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
gspread==4.0.1
openai==0.27.0
oauth2client==4.1.3
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.20.0
